+++ lib/src/framework/chart/indicator.dart
@@ -29,7 +29,6 @@ import '../../extension/export.dart';
 import '../../model/export.dart';
 import '../../utils/algorithm_util.dart';
 import '../../utils/platform_util.dart';
-import '../../indicators/candle/candle.dart';
 import '../collection/fixed_hash_queue.dart';
 import '../collection/sortable_hash_set.dart';
 import '../configuration.dart';